Add article on the BCH development fund proposal
Currently, there is a lot of debate happening over adding a development/
mining fee to the Bitcoin Cash node software. A lot of people are throw-
ing around the word coercive because it takes away some of their current
profits. Others have concerns about centralization.

I don't believe the mining fee is coercive at all. However, I am concer-
ned about the centralizing effect of a development fund. This article
sets out to establish that the mining fee is not coercive, and suggest
some updates to the proposal to avoid the centralization effects.